requirements.txt
beautifulsoup4==4.4.1
bleach==1.4.2
chardet==2.3.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-ckeditor==5.0.2
django-contrib-comments==1.6.1
filebrowser-safe==0.4.0
future==0.15.2
grappelli-safe==0.4.1
gunicorn==19.3.0
html5lib==0.9999999
Mezzanine==4.0.1
oauthlib==1.0.3
Pillow==3.0.0
pytz==2015.6
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
static3==0.6.1
tzlocal==1.2
